@@ -1,19 +1,21 @@
 {
-  "name": "yc-task1",
+  "name": "shiptivitas",
   "version": "0.1.0",
-  "engines": {
-    "npm": "6.7.0",
-    "node": "11.10.0"
-  },
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.2.1",
-    "dragula": "^3.7.2",
-    "jquery": "^3.3.1",
-    "popper.js": "^1.14.7",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-scripts": "2.1.3"
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "bootstrap": "^5.3.2",
+    "dragula": "^3.7.3",
+    "jquery": "^3.7.1",
+    "popper": "^1.0.1",
+    "react": "^18.2.0",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-dom": "^18.2.0",
+    "react-dragula": "^1.1.17",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -22,12 +24,24 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ],
+    "permissions": [
+      "http://localhost:3000/"
+    ]
+  }
 }
