@@ -1,12 +1,10 @@
-import React from 'react';
+import React from "react";
 
 export default function HomeTab() {
   return (
     <header className="App-header">
       <div className="App-title">
-        <h1>
-          SHIPTIVITAS
-        </h1>
+        <h1>SHIPTIVITAS</h1>
         <div className="App-icon">
           <div className="Circle"></div>
         </div>
